Add JsonParser class
